Create abstract base class to relate CompositionHighlight and CompositionUnderline https://bugs.webkit.org/show_bug.cgi?id=257350 rdar://109853785
Reviewed by NOBODY (OOPS!).

Trying to expand either type or create a third, similar type is difficult because these types are implemented separately. Implementing them with a common ancestor will make it easier to extend them in the future to support more use cases.

* Source/WebCore/WebCore.xcodeproj/project.pbxproj:
* Source/WebCore/editing/CompositionAttribute.h: Renamed from Source/WebCore/editing/CompositionHighlight.h.
(WebCore::CompositionAttribute::CompositionAttribute):
(WebCore::CompositionAttribute::operator<=> const):
(WebCore::CompositionHighlight::CompositionHighlight):
(WebCore::CompositionUnderline::CompositionUnderline):
* Source/WebCore/editing/CompositionUnderline.h: Removed.
* Source/WebCore/editing/Editor.cpp:
* Source/WebCore/editing/Editor.h:
* Source/WebCore/platform/PlatformKeyboardEvent.h:
* Source/WebCore/rendering/LegacyInlineTextBox.cpp:
* Source/WebCore/rendering/TextBoxPainter.cpp:
* Source/WebKit/Scripts/webkit/messages.py:
(headers_for_type):
* Source/WebKit/Shared/WebCoreArgumentCoders.cpp:
* Source/WebKit/Shared/WebCoreArgumentCoders.serialization.in:
* Source/WebKit/Shared/WebKeyboardEvent.h:
* Source/WebKit/UIProcess/WebPageProxy.cpp:
* Source/WebKit/UIProcess/ios/WKContentViewInteraction.mm:
* Source/WebKit/UIProcess/mac/WebViewImpl.mm:
* Source/WebKit/WebProcess/InjectedBundle/API/c/WKBundlePage.cpp:
* Source/WebKitLegacy/mac/WebView/WebFrame.mm:
* Source/WebKitLegacy/mac/WebView/WebHTMLView.mm:
