+++ Source/WebCore/editing/Editor.cpp
@@ -37,8 +37,8 @@
 #include "CachedResourceLoader.h"
 #include "ChangeListTypeCommand.h"
 #include "ClipboardEvent.h"
+#include "CompositionAttribute.h"
 #include "CompositionEvent.h"
-#include "CompositionHighlight.h"
 #include "CreateLinkCommand.h"
 #include "CustomUndoStep.h"
 #include "DataTransfer.h"
