+++ Source/WebKit/WebProcess/InjectedBundle/API/c/WKBundlePage.cpp
@@ -61,7 +61,7 @@
 #include <WebCore/AccessibilityObjectInterface.h>
 #include <WebCore/ApplicationCacheStorage.h>
 #include <WebCore/CSSParser.h>
-#include <WebCore/CompositionHighlight.h>
+#include <WebCore/CompositionAttribute.h>
 #include <WebCore/FocusController.h>
 #include <WebCore/LocalFrame.h>
 #include <WebCore/Page.h>
